Rename spidermonkey.py to jsshell.py, to avoid conflicts with the spidermonkey package.
